RecView pinch animation, optimized image loading, display selected amount

+++ app/src/main/java/com/penguinstudio/safecrypt/models/MediaModel.kt
@@ -5,7 +5,7 @@ enum class MediaType {
     IMAGE, VIDEO
 }
 data class MediaModel(
-    var id: Int,
+    var id: Long,
     val mediaUri: Uri,
     var albumName: String?,
     var mediaType: MediaType?,
